requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
discord-webhook>=1.3.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
